+++ src/get-css.js
@@ -21,6 +21,8 @@ import getUrls from './get-urls';
  * @param {object}   [options.filter] Regular expression used to filter node CSS
  *                   data. Each block of CSS data is tested against the filter,
  *                   and only matching data is included.
+ * @param {boolean}  [options.skipDisabled=true] Determines if disabled
+ *                   stylesheets will be skipped while collecting CSS data.
  * @param {boolean}  [options.useCSSOM=false] Determines if CSS data will be
  *                   collected from a stylesheet's runtime values instead of its
  *                   text content. This is required to get accurate CSS data
@@ -49,6 +51,7 @@ import getUrls from './get-urls';
  *     include     : 'style,link[rel="stylesheet"]',
  *     exclude     : '[href="skip.css"]',
  *     filter      : /red/,
+ *     skipDisabled: true,
  *     useCSSOM    : false,
  *     onBeforeSend(xhr, node, url) {
  *       // ...
@@ -76,6 +79,7 @@ function getCssData(options) {
         include     : options.include      || 'style,link[rel="stylesheet"]',
         exclude     : options.exclude      || null,
         filter      : options.filter       || null,
+        skipDisabled: options.skipDisabled !== false,
         useCSSOM    : options.useCSSOM     || false,
         onBeforeSend: options.onBeforeSend || Function.prototype,
         onSuccess   : options.onSuccess    || Function.prototype,
@@ -242,10 +246,11 @@ function getCssData(options) {
         sourceNodes.forEach((node, i) => {
             const linkHref = node.getAttribute('href');
             const linkRel  = node.getAttribute('rel');
-            const isLink   = node.nodeName === 'LINK' && linkHref && linkRel && linkRel.toLowerCase() === 'stylesheet';
+            const isLink   = node.nodeName === 'LINK' && linkHref && linkRel && linkRel.toLowerCase().indexOf('stylesheet') !== -1;
+            const isSkip   = settings.skipDisabled === false ? false : node.disabled;
             const isStyle  = node.nodeName === 'STYLE';
 
-            if (isLink) {
+            if (isLink && !isSkip) {
                 getUrls(linkHref, {
                     mimeType: 'text/css',
                     onBeforeSend(xhr, url, urlIndex) {
@@ -264,7 +269,7 @@ function getCssData(options) {
                     }
                 });
             }
-            else if (isStyle) {
+            else if (isStyle && !isSkip) {
                 let cssText = node.textContent;
 
                 if (settings.useCSSOM) {